# [WIP] Optimize varlena output functions