--- a/liblwgeom/cunit/cu_out_kml.c
+++ b/liblwgeom/cunit/cu_out_kml.c
@@ -19,58 +19,38 @@
 
 static void do_kml_test(char * in, char * out, int precision)
 {
-	LWGEOM *g;
-	char * h;
+	LWGEOM *g = lwgeom_from_wkt(in, LW_PARSER_CHECK_NONE);
+	lwvarlena_t *v = lwgeom_to_kml2(g, precision, "");
 
-	g = lwgeom_from_wkt(in, LW_PARSER_CHECK_NONE);
-	h = lwgeom_to_kml2(g, precision, "");
-
-	if (strcmp(h, out))
-		fprintf(stderr, "\nIn:   %s\nOut:  %s\nTheo: %s\n", in, h, out);
-
-	CU_ASSERT_STRING_EQUAL(h, out);
+	ASSERT_VARLENA_EQUAL(v, out);
 
 	lwgeom_free(g);
-	lwfree(h);
+	lwfree(v);
 }
 
 
 static void do_kml_unsupported(char * in, char * out)
 {
-	LWGEOM *g;
-	char *h;
+	LWGEOM *g = lwgeom_from_wkt(in, LW_PARSER_CHECK_NONE);
+	lwvarlena_t *v = lwgeom_to_kml2(g, 0, "");
 
-	g = lwgeom_from_wkt(in, LW_PARSER_CHECK_NONE);
-	h = lwgeom_to_kml2(g, 0, "");
-
-	if (strcmp(cu_error_msg, out))
-		fprintf(stderr, "\nIn:   %s\nOut:  %s\nTheo: %s\n",
-		        in, cu_error_msg, out);
-
-	CU_ASSERT_STRING_EQUAL(out, cu_error_msg);
+	ASSERT_STRING_EQUAL(cu_error_msg, out);
 	cu_error_msg_reset();
 
-	lwfree(h);
+	lwfree(v);
 	lwgeom_free(g);
 }
 
 
 static void do_kml_test_prefix(char * in, char * out, int precision, const char *prefix)
 {
-	LWGEOM *g;
-	char * h;
+	LWGEOM *g = lwgeom_from_wkt(in, LW_PARSER_CHECK_NONE);
+	lwvarlena_t *v = lwgeom_to_kml2(g, precision, prefix);
 
-	g = lwgeom_from_wkt(in, LW_PARSER_CHECK_NONE);
-	h = lwgeom_to_kml2(g, precision, prefix);
-
-	if (strcmp(h, out))
-		fprintf(stderr, "\nPrefix: %s\nIn:   %s\nOut:  %s\nTheo: %s\n",
-		        prefix, in, h, out);
-
-	CU_ASSERT_STRING_EQUAL(h, out);
+	ASSERT_VARLENA_EQUAL(v, out);
 
 	lwgeom_free(g);
-	lwfree(h);
+	lwfree(v);
 }