… liblwgeom/cunit/cu_surface.c
… liblwgeom/gserialized.txt
… liblwgeom/gserialized2.c
… liblwgeom/liblwgeom_internal.h
… liblwgeom/lwalgorithm.c
… liblwgeom/lwout_encoded_polyline.c
… liblwgeom/lwout_geojson.c
… liblwgeom/lwout_gml.c
… liblwgeom/lwout_kml.c
… liblwgeom/lwout_svg.c
… liblwgeom/lwout_twkb.c
… liblwgeom/lwout_wkb.c
… liblwgeom/lwout_wkt.c
… liblwgeom/lwout_x3d.c
… liblwgeom/stringbuffer.c
… liblwgeom/stringbuffer.h
… loader/shp2pgsql-core.c
… postgis/geography_inout.c
… postgis/lwgeom_dump.c
… postgis/lwgeom_export.c
… postgis/lwgeom_functions_basic.c
… postgis/lwgeom_geos.c
… postgis/lwgeom_inout.c
… postgis/lwgeom_ogc.c
… postgis/lwgeom_transform.c
… raster/rt_core/rt_geometry.c
… raster/rt_pg/rtpg_geometry.c
… raster/rt_pg/rtpg_mapalgebra.c
… raster/rt_pg/rtpg_pixel.c
… topology/postgis_topology.c