Related to https://trac.osgeo.org/postgis/ticket/4623

The idea behind this PR is to enable liblwgeom to output an array that could be returned directly to Postgresql without any copies. I want to understand how invasive the changes are and what's the impact in performance.

This could potentially be used in multiple output functions such as `ST_AsGeoJSON` (first commit here), `ST_AsBinary`, `ST_AsText`...