@Algunenano 2020-02-14
Tested ST_AsText and the impact is basically nothing. I'm going to test ST_AsBinary and ST_AsTWKB just in case; ST_AsBinary at least it's faster, so the impact of not copying buffers should be more noticeable. If it isn't, I'll discard this completely.