@Algunenano 2020-03-13
I've tested (https://github.com/Algunenano/postgis/tree/varlena_wkb) an implementation of ST_AsBinary (and all associated functions) and it is 6-22% faster but I have introduced some bug at some point since I have some failures. It does look promising so I'll review everything with more calm to introduce a commit here with the changes.