@Algunenano 2020-03-20
Implemented over the WKB variants (there are multiple functions affected). Measurements:

```
-- 683788 Points (1 vertex on avg :D)ls
-- EXPLAIN ANALYZE Select ST_AsBinary(the_geom) from benchmark_7773a711c8441d4b494a51fd9feebeac7a9b9c734619398620293;
-- Before: 7.807002 tps
-- After: 8.635077 tps
-- Perf: 1.1060x

-- 489298 Lines (~5.7 vertices on avg)
-- EXPLAIN ANALYZE Select ST_AsBinary(the_geom) from benchmark_da625777246e392648f7d78b912233019d0cb2ffbf91acb187913;
-- Before: 7.621094 tps
-- After: 8.075582 tps
-- Perf: 1.0596x

-- 1083175 Small polygons (~9 vertices on avg)
-- EXPLAIN ANALYZE Select ST_AsBinary(the_geom) from benchmark_af85af967525e69373b7b8678f3bd409a8e78f447671344b7535d;
-- Before: 3.841667 tps
-- After: 4.216442 tps
-- Perf: 1.0975x

-- 13 big polygons (~260843 vertices on avg)
-- EXPLAIN ANALYZE Select ST_AsBinary(the_geom) from benchmark_4c7214d90a79aa6760367a084a4d4a2f61fbe1c6cc4f7f9e76020;
-- Before: 20.638837 tps
-- After: 24.819581 tps
-- Perf: 1.2025x
```