@Algunenano 2020-04-03
TWKB changed too. The improvements are marginal (1-3%):
```
-- 683788 Points (1 vertex on avg)
-- EXPLAIN ANALYZE Select ST_AsTWKB(the_geom) from benchmark_7773a711c8441d4b494a51fd9feebeac7a9b9c734619398620293;
-- Before: 7.094080 tps
-- After: 7.143945 tps
-- Perf: 1.0070x

-- 489298 Lines (~5.7 vertices on avg)
-- EXPLAIN ANALYZE Select ST_AsTWKB(the_geom) from benchmark_da625777246e392648f7d78b912233019d0cb2ffbf91acb187913;
-- Before: 6.783711 tps
-- After: 6.853756 tps
-- Perf: 1.0103x

-- 1083175 Small polygons (~9 vertices on avg)
-- EXPLAIN ANALYZE Select ST_AsTWKB(the_geom) from benchmark_af85af967525e69373b7b8678f3bd409a8e78f447671344b7535d;
-- Before: 3.495658 tps
-- After: 3.550130 tps
-- Perf: 1.0155x

-- 13 big polygons (~260843 vertices on avg)
EXPLAIN ANALYZE Select ST_AsTWKB(the_geom) from benchmark_4c7214d90a79aa6760367a084a4d4a2f61fbe1c6cc4f7f9e76020;
-- Before: 11.073627 tps
-- After: 11.379276 tps
-- Perf: 1.0276x
```

I'm going to review some pending output functions to see if they can also benefit from this kind of changes (I would expect so): KML, GML, SVG, X3D, GeoHash