@Algunenano 2020-04-03
Also a marginal improvement (1-2%) for ST_AsKML:
```
-- 683788 Points (1 vertex on avg)
EXPLAIN ANALYZE Select ST_AsKML(the_geom) from benchmark_7773a711c8441d4b494a51fd9feebeac7a9b9c734619398620293;
-- Before: 2.527324 tps
-- After: 2.544996 tps
-- Perf: 1.0069 x

-- 489298 Lines (~5.7 vertices on avg)
-- EXPLAIN ANALYZE Select ST_AsKML(the_geom) from benchmark_da625777246e392648f7d78b912233019d0cb2ffbf91acb187913;
-- Before: 0.707353 tps
-- After: 0.714354 tps
-- Perf: 1.0098 x

-- 1083175 Small polygons (~9 vertices on avg)
-- EXPLAIN ANALYZE Select ST_AsKML(the_geom) from benchmark_af85af967525e69373b7b8678f3bd409a8e78f447671344b7535d;
-- Before: 0.239224 tps
-- After: 0.241276 tps
-- Perf: 1.0085 x

-- 13 big polygons (~260843 vertices on avg)
-- EXPLAIN ANALYZE Select ST_AsKML(the_geom) from benchmark_4c7214d90a79aa6760367a084a4d4a2f61fbe1c6cc4f7f9e76020;
-- Before: 0.178417 tps
-- After: 0.181457 tps
-- Perf: 1.0170 x
```