@Algunenano 2020-04-03
Implemented for GeoHash, ~10% improvements for points (I haven't tested other cases, since I don't think are significative):
```
-- 683788 Points (1 vertex on avg)
EXPLAIN ANALYZE Select ST_GeoHash(the_geom) from benchmark_7773a711c8441d4b494a51fd9feebeac7a9b9c734619398620293;
-- Before: 4.855698 tps
-- After: 5.299335 tps
-- Perf: 1.091 x
```