@Algunenano 2020-04-22
I think that now covers every important output function. I'll do some final benchmarks and merge unless anybody complains.