@Algunenano 2020-04-22
X3D: ~3% improvement
```
-- 683788 Points (1 vertex on avg)
EXPLAIN ANALYZE Select ST_AsX3D(the_geom) from benchmark_7773a711c8441d4b494a51fd9feebeac7a9b9c734619398620293;
-- Before: 0.025865 tps
-- After: 0.026718 tps
-- Perf: 1.0329x

-- 489298 Lines (~5.7 vertices on avg)
-- EXPLAIN ANALYZE Select ST_AsX3D(the_geom) from benchmark_da625777246e392648f7d78b912233019d0cb2ffbf91acb187913;
-- Before: 0.033178 tps
-- After: 0.034265 tps
-- Perf: 1.0327 x

-- 1083175 Small polygons (~9 vertices on avg)
-- EXPLAIN ANALYZE Select ST_AsX3D(the_geom) from benchmark_af85af967525e69373b7b8678f3bd409a8e78f447671344b7535d;
-- Before: 0.014596 tps
-- After: 0.015085 tps
-- Perf:  1.0335x

-- 13 big polygons (~260843 vertices on avg)
-- EXPLAIN ANALYZE Select ST_AsX3D(the_geom) from benchmark_4c7214d90a79aa6760367a084a4d4a2f61fbe1c6cc4f7f9e76020;
-- Before: 0.713027 tps
-- After: 0.738587 tps
-- Perf: 1.0358x
```

ST_AsGML: 0 to 7% improvement
```
-- 683788 Points (1 vertex on avg)
EXPLAIN ANALYZE Select ST_AsGML(the_geom) from benchmark_7773a711c8441d4b494a51fd9feebeac7a9b9c734619398620293;
-- Before: 0.119722 tps
-- After: 0.120075 tps
-- Perf: 1.0029 x

-- 489298 Lines (~5.7 vertices on avg)
-- EXPLAIN ANALYZE Select ST_AsGML(the_geom) from benchmark_da625777246e392648f7d78b912233019d0cb2ffbf91acb187913;
-- Before: 0.124858 tps
-- After: 0.126692 tps
-- Perf: 1.0146 x

-- 1083175 Small polygons (~9 vertices on avg)
-- EXPLAIN ANALYZE Select ST_AsGML(the_geom) from benchmark_af85af967525e69373b7b8678f3bd409a8e78f447671344b7535d;
-- Before: 0.069138 tps
-- After: 0.069964 tps
-- Perf: 1.0119 x

-- 13 big polygons (~260843 vertices on avg)
-- EXPLAIN ANALYZE Select ST_AsGML(the_geom) from benchmark_4c7214d90a79aa6760367a084a4d4a2f61fbe1c6cc4f7f9e76020;
-- Before: 0.813557 tps
-- After: 0.872248 tps
-- Perf: 1.0721 x
```

ST_AsSVG: 3 to 7% improvement:
```
-- 683788 Points (1 vertex on avg)
EXPLAIN ANALYZE Select ST_AsSVG(the_geom) from benchmark_7773a711c8441d4b494a51fd9feebeac7a9b9c734619398620293;
-- Before: 5.940995 tps
-- After: 6.125753 tps
-- Perf: 1.03109 x

-- 489298 Lines (~5.7 vertices on avg)
-- EXPLAIN ANALYZE Select ST_AsSVG(the_geom) from benchmark_da625777246e392648f7d78b912233019d0cb2ffbf91acb187913;
-- Before: 2.828125 tps
-- After: 2.916545 tps
-- Perf: 1.0312 x

-- 1083175 Small polygons (~9 vertices on avg)
-- EXPLAIN ANALYZE Select ST_AsSVG(the_geom) from benchmark_af85af967525e69373b7b8678f3bd409a8e78f447671344b7535d;
-- Before: 1.287415 tps
-- After: 1.326895 tps
-- Perf: 1.0306 x

-- 13 big polygons (~260843 vertices on avg)
-- EXPLAIN ANALYZE Select ST_AsSVG(the_geom) from benchmark_4c7214d90a79aa6760367a084a4d4a2f61fbe1c6cc4f7f9e76020;
-- Before: 0.932382 tps
-- After: 0.992763 tps
-- Perf: 1.0647 x
```

ST_AsEncodedPolyline: ~3% improvement
```
-- 489298 Lines (~5.7 vertices on avg)
EXPLAIN ANALYZE Select ST_AsEncodedPolyline(ST_LineMerge(the_geom)) from benchmark_da625777246e392648f7d78b912233019d0cb2ffbf91acb187913;
-- Before: 1.640634 tps
-- After: 1.692809 tps
-- Perf: 1.0318 x
```